requirements.txt
pyhomebroker
requests
websocket-client